Introduces several optimizations to reduce memory consumption for dcor benchmark
The dcor benchmark is based on actual use of renjin that is currently maxing out heap
space and running too slowly. The algorithm is written in idiomatic R and results in big
copies of data all over the place. 

This commit introduces a few optimizations:
- For large matrices, aperm() returns a wrapper rather than actually creating a transposed
copy of the matrix. (Probably want to do this in t.default() as well)
- StringArrayVector has been extracted from StringVector in the same way we've done 
for IntVector, DoubleVector, etc, making it possible to avoid allocation of giant name vectors
- For large vectors, as.character(x) simply wraps the underlying vector and converts 
the elements to strings as needed. So the expression as.character(1:10000000) now actually
allocates no memory and runs in constant time. This is particularly useful since the names
vector is often defined without actually being used

More to come...

git-svn-id: https://renjin.googlecode.com/svn/trunk@621 b722d886-ae41-2056-5ac6-6db9963f4c99